requirements.txt
keras==2.4.3
tensorflow==2.4.0
Pillow
numpy~=1.19.4